+++ addons/gut/gui/gut_config_gui.gd
@@ -23,6 +23,7 @@ class DirectoryCtrl:
 	func _on_selected(path):
 		set_text(path)
 
+
 	func _on_dir_button_pressed():
 		_dialog.current_dir = _txt_path.text
 		_dialog.popup_centered()
@@ -40,6 +41,9 @@ class DirectoryCtrl:
 		text = t
 		_txt_path.text = text
 
+	func get_line_edit():
+		return _txt_path
+
 # ------------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
 class FileCtrl:
@@ -47,6 +51,50 @@ class FileCtrl:
 
 	func _init():
 		_dialog.mode = _dialog.MODE_OPEN_FILE
+
+# ------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
+class Vector2Ctrl:
+	extends VBoxContainer
+
+	var value = Vector2(-1, -1) setget set_value, get_value
+	var disabled = false setget set_disabled, get_disabled
+	var x_spin = SpinBox.new()
+	var y_spin = SpinBox.new()
+
+	func _init():
+		add_child(_make_one('x:  ', x_spin))
+		add_child(_make_one('y:  ', y_spin))
+
+	func _make_one(txt, spinner):
+		var hbox = HBoxContainer.new()
+		var lbl = Label.new()
+		lbl.text = txt
+		hbox.add_child(lbl)
+		hbox.add_child(spinner)
+		spinner.min_value = -1
+		spinner.max_value = 10000
+		spinner.size_flags_horizontal = spinner.SIZE_EXPAND_FILL
+		return hbox
+
+	func set_value(v):
+		if(v != null):
+			x_spin.value = v[0]
+			y_spin.value = v[1]
+
+	# Returns array instead of vector2 b/c that is what is stored in
+	# in the dictionary and what is expected everywhere else.
+	func get_value():
+		return [x_spin.value, y_spin.value]
+
+	func set_disabled(should):
+		get_parent().visible = !should
+		x_spin.visible = !should
+		y_spin.visible = !should
+
+	func get_disabled():
+		pass
+
 
 
 # ------------------------------------------------------------------------------
@@ -90,6 +138,10 @@ func _new_row(key, disp_text, value_ctrl, hint):
 	_cfg_ctrls[key] = value_ctrl
 	ctrl.add_child(value_ctrl)
 
+	var rpad = CenterContainer.new()
+	rpad.rect_min_size.x = 5
+	ctrl.add_child(rpad)
+
 	return ctrl
 
 
@@ -107,8 +159,10 @@ func _add_title(text):
 func _add_number(key, value, disp_text, v_min, v_max, hint=''):
 	var value_ctrl = SpinBox.new()
 	value_ctrl.value = value
+	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
 	value_ctrl.min_value = v_min
 	value_ctrl.max_value = v_max
+	_wire_select_on_focus(value_ctrl.get_line_edit())
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
@@ -121,6 +175,7 @@ func _add_select(key, value, values, disp_text, hint=''):
 		if(value == values[i]):
 			select_idx = i
 	value_ctrl.selected = select_idx
+	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
@@ -129,6 +184,7 @@ func _add_value(key, value, disp_text, hint=''):
 	var value_ctrl = LineEdit.new()
 	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
 	value_ctrl.text = value
+	_wire_select_on_focus(value_ctrl)
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
@@ -139,10 +195,12 @@ func _add_boolean(key, value, disp_text, hint=''):
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
+
 func _add_directory(key, value, disp_text, hint=''):
 	var value_ctrl = DirectoryCtrl.new()
 	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
 	value_ctrl.text = value
+	_wire_select_on_focus(value_ctrl.get_line_edit())
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
@@ -151,6 +209,7 @@ func _add_file(key, value, disp_text, hint=''):
 	var value_ctrl = FileCtrl.new()
 	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
 	value_ctrl.text = value
+	_wire_select_on_focus(value_ctrl.get_line_edit())
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
 
@@ -161,11 +220,54 @@ func _add_color(key, value, disp_text, hint=''):
 	value_ctrl.color = value
 
 	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
+
+
+func _add_vector2(key, value, disp_text, hint=''):
+	var value_ctrl = Vector2Ctrl.new()
+	value_ctrl.size_flags_horizontal = value_ctrl.SIZE_EXPAND_FILL
+	value_ctrl.value = value
+	_wire_select_on_focus(value_ctrl.x_spin.get_line_edit())
+	_wire_select_on_focus(value_ctrl.y_spin.get_line_edit())
+
+	var ctrl = _new_row(key, disp_text, value_ctrl, hint)
+# -----------------------------
+
+
+func _wire_up_controls():
+	_cfg_ctrls.use_viewport.connect('toggled', self, '_on_use_viewport_toggled')
+
+
+func _show_viewport_controls(should):
+	_cfg_ctrls.resolution.disabled = !should
+	_cfg_ctrls.viewport_size.disabled = !should
+	_cfg_ctrls.gut_on_top.get_parent().visible = should
+	_cfg_ctrls.viewport_bg_color.get_parent().visible = should
+
+
 # ------------------
 # Events
 # ------------------
+func _on_use_viewport_toggled(pressed):
+	_show_viewport_controls(pressed)
+
+
 func _on_ctrl_value_changed(which):
 	pass
+
+
+func _wire_select_on_focus(which):
+	which.connect('focus_entered', self, '_on_ctrl_focus_highlight', [which])
+	which.connect('focus_exited', self, '_on_ctrl_focus_unhighlight', [which])
+
+
+func _on_ctrl_focus_highlight(which):
+	if(which.has_method('select_all')):
+		which.call_deferred('select_all')
+
+
+func _on_ctrl_focus_unhighlight(which):
+	if(which.has_method('select')):
+		which.select(0, 0)
 
 
 func _on_title_cell_draw(which):
@@ -215,12 +317,6 @@ func set_options(options):
 		"Exit if there are no failures.  Does nothing if 'Exit on Finish' is enabled.")
 
 
-	_add_title("XML Output")
-	_add_value("junit_xml_file", options.junit_xml_file, "Output Path",
-		"Path and filename where GUT should create the JUnit XML file.")
-	_add_boolean("junit_xml_timestamp", options.junit_xml_timestamp, "Include timestamp",
-		"Include a timestamp in the filename so that each run gets its own xml file.")
-
 	_add_title("Panel Output")
 	_add_select('output_font_name', options.panel_options.font_name, _avail_fonts, 'Font',
 		"The name of the font to use when running tests and in the output panel to the left.")
@@ -230,20 +326,41 @@ func set_options(options):
 
 	_add_title('Runner Appearance')
 	_add_select('font_name', options.font_name, _avail_fonts, 'Font',
-		"The name of the font to use when running tests and in the output panel to the left.")
+		"The font to use for text output in the Gut Runner.")
 	_add_number('font_size', options.font_size, 'Font Size', 5, 100,
-		"The font size to use when running tests and in the output panel to the left.")
+		"The font size for text output in the Gut Runner.")
+	_add_color('font_color', options.font_color, 'Font Color',
+		"The font color for text output in the Gut Runner.")
+	_add_color('background_color', options.background_color, 'Background Color',
+		"The background color for text output in the Gut Runner.")
 	_add_boolean('should_maximize', options.should_maximize, 'Maximize',
 		"Maximize GUT when tests are being run.")
 	_add_number('opacity', options.opacity, 'Opacity', 0, 100,
 		"The opacity of GUT when tests are running.")
-	_add_color('background_color', options.background_color, 'Background Color')
-	_add_color('font_color', options.font_color, 'Font Color')
 	_add_boolean('disable_colors', options.disable_colors, 'Disable Formatting',
 		'Disable formatting and colors used in the Runner.  Does not affect panel output.')
 
 
-	_add_title('Directories')
+	_add_title("Runner Window")
+	_add_boolean("use_viewport", options.use_viewport, "Use Viewport",
+		"Tests are added to the tree in a viewport instead of to the GUT " + \
+		"instance.  This can help with layout while tests are running")
+	_add_boolean("gut_on_top", options.gut_on_top, "GUT on Top",
+		"The GUT Runner appears above the viewport.")
+	_add_vector2('resolution', options.resolution, 'Resolution',
+		"The size and resolution of the window when tests are run.  Set both " +
+		"the x and y to non-zero values to override the project setting size.  " +
+		"The viewport retains the size of the project settings window size " +
+		"unless Viewport Size is set.")
+	_add_vector2('viewport_size', options.viewport_size, 'Viewport Size',
+		"Change the size of the viewport.  The viewport defaults to the project settings " +
+		"window size.")
+	_add_color('viewport_bg_color', options.viewport_bg_color, 'Viewport Bg Color',
+		"The color shown behind the viewport.  This color will bleed through " +
+		"transparencies of objects when you use add_child in a test.")
+
+
+	_add_title('Test Directories')
 	_add_boolean('include_subdirs', options.include_subdirs, 'Include Subdirs',
 		"Include subdirectories of the directories configured below.")
 	for i in range(DIRS_TO_LIST):
@@ -253,15 +370,28 @@ func set_options(options):
 
 		_add_directory(str('directory_', i), value, str('Directory ', i))
 
+	_add_title("XML Output")
+	_add_value("junit_xml_file", options.junit_xml_file, "Output Path",
+		"Path and filename where GUT should create a JUnit compliant XML file.  " +
+		"This file will contain the results of the last test run.  To avoid " +
+		"overriding the file use Include Timestamp.")
+	_add_boolean("junit_xml_timestamp", options.junit_xml_timestamp, "Include Timestamp",
+		"Include a timestamp in the filename so that each run gets its own xml file.")
+
 
 	_add_title('Hooks')
-	_add_file('pre_run_script', options.pre_run_script, 'Pre-Run Hook', 'This script will be run by GUT before any tests are run.')
-	_add_file('post_run_script', options.post_run_script, 'Post-Run Hook', 'This script will be run by GUT after all tests are run.')
+	_add_file('pre_run_script', options.pre_run_script, 'Pre-Run Hook',
+		'This script will be run by GUT before any tests are run.')
+	_add_file('post_run_script', options.post_run_script, 'Post-Run Hook',
+		'This script will be run by GUT after all tests are run.')
 
 
 	_add_title('Misc')
 	_add_value('prefix', options.prefix, 'Script Prefix',
 		"The filename prefix for all test scripts.")
+
+	_show_viewport_controls(options.use_viewport)
+	_wire_up_controls()
 
 
 func get_options(base_opts):
@@ -273,10 +403,6 @@ func get_options(base_opts):
 	to_return.hide_orphans = _cfg_ctrls.hide_orphans.pressed
 	to_return.should_exit = _cfg_ctrls.should_exit.pressed
 	to_return.should_exit_on_success = _cfg_ctrls.should_exit_on_success.pressed
-
-	# XML Output
-	to_return.junit_xml_file = _cfg_ctrls.junit_xml_file.text
-	to_return.junit_xml_timestamp = _cfg_ctrls.junit_xml_timestamp.pressed
 
 	#Output
 	to_return.panel_options.font_name = _cfg_ctrls.output_font_name.get_item_text(
@@ -293,6 +419,12 @@ func get_options(base_opts):
 	to_return.font_color = _cfg_ctrls.font_color.color.to_html()
 	to_return.disable_colors = _cfg_ctrls.disable_colors.pressed
 
+	# Test Window
+	to_return.use_viewport = _cfg_ctrls.use_viewport.pressed
+	to_return.resolution = _cfg_ctrls.resolution.value
+	to_return.viewport_size = _cfg_ctrls.viewport_size.value
+	to_return.gut_on_top = _cfg_ctrls.gut_on_top.pressed
+	to_return.viewport_bg_color = _cfg_ctrls.viewport_bg_color.color.to_html()
 
 	# Directories
 	to_return.include_subdirs = _cfg_ctrls.include_subdirs.pressed
@@ -303,6 +435,10 @@ func get_options(base_opts):
 		if(val != '' and val != null):
 			dirs.append(val)
 	to_return.dirs = dirs
+
+	# XML Output
+	to_return.junit_xml_file = _cfg_ctrls.junit_xml_file.text
+	to_return.junit_xml_timestamp = _cfg_ctrls.junit_xml_timestamp.pressed
 
 	# Hooks
 	to_return.pre_run_script = _cfg_ctrls.pre_run_script.text