+++ test/unit/test_gut.gd
@@ -146,6 +146,9 @@ func test_get_set_junit_xml_file():
 
 func test_get_set_junit_xml_timestamp():
 	assert_accessors(gr.test_gut, 'junit_xml_timestamp', false, true)
+
+func test_get_set_add_children_to():
+	assert_accessors(gr.test_gut, 'add_children_to', gr.test_gut, autofree(Node.new()))
 # ------------------------------
 # Doubler
 # ------------------------------
@@ -165,6 +168,7 @@ func test_clears_ignored_methods_between_tests():
 	gr.test_gut._tests_like = 'test_assert_eq_number_not_equal'
 	gr.test_gut.test_scripts()
 	assert_eq(gr.test_gut.get_doubler().get_ignored_methods().size(), 0)
+	pause_before_teardown()
 
 
 # ------------------------------